requirements.txt
Django==5.1.7
djangorestframework==3.14.0
openpyxl==3.1.5
pandas==2.2.3